+++ tools/libexec/perl/TTP/SqlServer.pm
@@ -702,6 +702,7 @@ sub restoreDatabase {
 # (I):
 # - the TTP EP entry point
 # - an argument object with following keys:
+#   > node: the TTP::Node object this DBMS runs on
 #   > service: the TTP::Service object this DBMS belongs to
 # (O):
 # - this object, or undef in case of an error
@@ -714,7 +715,7 @@ sub new {
 
 	if( $self ){
 		bless $self, $class;
-		msgVerbose( __PACKAGE__."::new()" );
+		msgVerbose( __PACKAGE__."::new() node='".$args->{node}->name()."' service='".$args->{service}->name()."'" );
 	}
 
 	return $self;